move account id to session
new plan is to work towards an implementation of openid connect with session
management. this calls for maintaining a session at the provider and checking
against it from the application.

the application will still need to know the account id, of course, but we'll
expose it using openid connect standards.