switch authn session to JWT
better consistency!

we use HMAC for the session signature and transport it over a secured cookie, like rails.
this takes care of the threat model and makes checking aud/iss a bit redundant. also, the
jwt doesn't expire because the token wrapped inside does. i'd rather not rewrite the authn
cookie continuously to match the updated RefreshToken TTL.

FUN:
* this JWT is only intended for AuthN, so it keeps track of the app via the `azp` claim.
* the derived key used to generate the session signature has many more iterations than Rails,
  and uses SHA-256 instead of SHA1.